Modified - move collect screen provider
