+++ lib/features/customer_feature/presentation/widgets/profile_view.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:provider/provider.dart';
-import 'package:tamizshahr/screens/collect_list_screen.dart';
+import 'package:tamizshahr/features/waste_feature/collect_list_screen.dart';
 import 'package:tamizshahr/features/store_feature/presentation/screens/orders_screen.dart';
 
 import '../../../../core/widgets/top_bar.dart';
