+++ lib/features/home_feature/presentation/home_screen.dart
@@ -8,7 +8,7 @@ import '../../../provider/Products.dart';
 import '../../../core/theme/app_theme.dart';
 import '../../customer_feature/presentation/providers/auth.dart';
 import '../../../screens/article_screen.dart';
-import '../../../screens/collect_list_screen.dart';
+import '../../waste_feature/collect_list_screen.dart';
 import '../../waste_feature/presentation/waste_cart_screen.dart';
 import '../../../core/widgets/custom_dialog.dart';
 import '../../store_feature/presentation/screens/product_screen.dart';
