+++ lib/features/waste_feature/collect_list_screen.dart
@@ -5,13 +5,13 @@ import 'package:provider/provider.dart';
 import 'package:tamizshahr/features/waste_feature/business/entities/request_waste_item.dart';
 import 'package:tamizshahr/features/customer_feature/presentation/providers/auth.dart';
 
-import '../models/search_detail.dart';
-import '../core/theme/app_theme.dart';
-import '../features/waste_feature/presentation/providers/wastes.dart';
-import '../widgets/collect_item_collect_screen.dart';
-import '../core/logic/en_to_ar_number_convertor.dart';
-import '../core/widgets/main_drawer.dart';
-import '../features/customer_feature/presentation/screens/login_screen.dart';
+import '../../models/search_detail.dart';
+import '../../core/theme/app_theme.dart';
+import 'presentation/providers/wastes.dart';
+import '../../widgets/collect_item_collect_screen.dart';
+import '../../core/logic/en_to_ar_number_convertor.dart';
+import '../../core/widgets/main_drawer.dart';
+import '../customer_feature/presentation/screens/login_screen.dart';
 
 class CollectListScreen extends StatefulWidget {
   static const routeName = '/collectListScreen';
