+++ lib/main.dart
@@ -7,7 +7,7 @@ import 'package:tamizshahr/provider/orders.dart';
 import 'package:tamizshahr/screens/charity_detail_screen.dart';
 import 'package:tamizshahr/screens/charity_screen.dart';
 import 'package:tamizshahr/screens/clear_screen.dart';
-import 'package:tamizshahr/screens/collect_detail_screen.dart';
+import 'package:tamizshahr/features/waste_feature/collect_detail_screen.dart';
 import 'package:tamizshahr/screens/donation_screen.dart';
 import 'package:tamizshahr/features/store_feature/presentation/screens/orders_screen.dart';
 import 'package:tamizshahr/screens/wallet_screen.dart';
@@ -38,7 +38,7 @@ import 'features/waste_feature/presentation/wastes_screen.dart';
 import 'core/constants/strings.dart';
 import 'provider/Products.dart';
 import 'features/customer_feature/presentation/providers/customer_info.dart';
-import 'screens/collect_list_screen.dart';
+import 'features/waste_feature/collect_list_screen.dart';
 import 'features/customer_feature/presentation/screens/customer_detail_info_edit_screen.dart';
 import 'features/customer_feature/presentation/screens/login_screen.dart';
 import 'features/customer_feature/presentation/screens/profile_screen.dart';
