+++ lib/widgets/collect_item_collect_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart' as intl;
 import 'package:provider/provider.dart';
 import 'package:tamizshahr/features/waste_feature/business/entities/request_waste_item.dart';
-import 'package:tamizshahr/screens/collect_detail_screen.dart';
+import 'package:tamizshahr/features/waste_feature/collect_detail_screen.dart';
 
 import '../provider/Products.dart';
 import '../core/theme/app_theme.dart';
